// src/model/PatternNames.ts
export type PatternNames =
  | "SINGLETON"
  | "FACTORY METHOD"
  | "ABSTRACT FACTORY"
  | "BUILDER"
  | "PROTOTYPE"
  | "ADAPTER"
  | "BRIDGE"
  | "COMPOSITE"
  | "DECORATOR"
  | "FACADE"
  | "FLYWEIGHT"
  | "PROXY"
  | "CHAIN OF RESPONSIBILITY"
  | "COMMAND"
  | "ITERATOR"
  | "MEDIATOR"
  | "MEMENTO"
  | "OBSERVER"
  | "STATE"
  | "STRATEGY"
  | "TEMPLATE METHOD"
  | "VISITOR"
  | "INTERPRETER";